+++ lib/main.js
@@ -1,10 +1,12 @@
+'use strict'
+
 const WritableStream = require('stream').Writable
 const { inherits } = require('util')
 const Dicer = require('../deps/dicer/lib/Dicer')
 
 const MultipartParser = require('./types/multipart')
 const UrlencodedParser = require('./types/urlencoded')
-const parseParams = require('./utils/decode').parseParams
+const parseParams = require('./utils/parseParams')
 
 function Busboy (opts) {
   if (!(this instanceof Busboy)) { return new Busboy(opts) }
