+++ lib/types/multipart.js
@@ -1,3 +1,5 @@
+'use strict'
+
 // TODO:
 //  * support 1 nested multipart level
 //    (see second multipart example here:
@@ -10,8 +12,8 @@ const inherits = require('util').inherits
 
 const Dicer = require('../../deps/dicer/lib/Dicer')
 
-const parseParams = require('../utils/decode').parseParams
-const decodeText = require('../utils/decode').decodeText
+const parseParams = require('../utils/parseParams')
+const decodeText = require('../utils/decodeText')
 const basename = require('../utils/basename')
 const getLimit = require('../utils/getLimit')
