+++ lib/types/urlencoded.js
@@ -1,5 +1,7 @@
-const Decoder = require('../utils/decode').Decoder
-const decodeText = require('../utils/decode').decodeText
+'use strict'
+
+const Decoder = require('../utils/Decoder')
+const decodeText = require('../utils/decodeText')
 const getLimit = require('../utils/getLimit')
 
 const RE_CHARSET = /^charset$/i
