+++ lib/utils/basename.js
@@ -1,3 +1,5 @@
+'use strict'
+
 module.exports = function basename (path) {
   if (typeof path !== 'string') { return '' }
   for (var i = path.length - 1; i >= 0; --i) { // eslint-disable-line no-var
