+++ lib/utils/getLimit.js
@@ -1,3 +1,5 @@
+'use strict'
+
 module.exports = function getLimit (limits, name, defaultLimit) {
   if (
     !limits ||
