+++ test/decoder.spec.js
@@ -1,5 +1,5 @@
 const { assert, expect } = require('chai')
-const { Decoder } = require('../lib/utils/decode')
+const Decoder = require('../lib/utils/Decoder')
 
 describe('Decoder', () => {
   [
